+++ src/components/Shared/PPM/Closeout/FinalCloseoutForm/FinalCloseoutForm.test.jsx
@@ -111,16 +111,6 @@ describe('FinalCloseoutForm component', () => {
     expect(defaultProps.onBack).toHaveBeenCalled();
   });
 
-  it('calls onBack func when "Return To Homepage" button is clicked', async () => {
-    const mtoShipment = createPPMShipmentWithFinalIncentive();
-
-    render(<FinalCloseoutForm mtoShipment={mtoShipment} {...defaultProps} />);
-
-    await userEvent.click(screen.getByRole('button', { name: 'Return To Homepage' }));
-
-    expect(defaultProps.onBack).toHaveBeenCalled();
-  });
-
   it('validates the form after user input', async () => {
     const mtoShipment = createPPMShipmentWithFinalIncentive();
     render(<FinalCloseoutForm mtoShipment={mtoShipment} {...defaultProps} />);
