+++ src/pages/Office/PPM/Closeout/FinalCloseout/FinalCloseout.jsx
@@ -6,13 +6,12 @@ import { Alert, Grid, GridContainer } from '@trussworks/react-uswds';
 import styles from './FinalCloseout.module.scss';
 
 import { useEditShipmentQueries } from 'hooks/queries';
-import FinalCloseoutForm from 'pages/Office/PPM/Closeout/FinalCloseoutForm/FinalCloseoutForm';
+import FinalCloseoutForm from 'components/Shared/PPM/Closeout/FinalCloseoutForm/FinalCloseoutForm';
 import NotificationScrollToTop from 'components/NotificationScrollToTop';
 import ShipmentTag from 'components/ShipmentTag/ShipmentTag';
 import { servicesCounselingRoutes } from 'constants/routes';
 import { shipmentTypes } from 'constants/shipments';
-import ppmPageStyles from 'pages/MyMove/PPM/PPM.module.scss';
-import { ppmSubmissionCertificationText } from 'scenes/Legalese/legaleseText';
+import ppmPageStyles from 'pages/Office/PPM/PPM.module.scss';
 import { getResponseError, submitPPMShipmentSignedCertification } from 'services/ghcApi';
 import LoadingPlaceholder from 'shared/LoadingPlaceholder';
 import SomethingWentWrong from 'shared/SomethingWentWrong';
@@ -37,17 +36,11 @@ const FinalCloseout = () => {
     navigate(generatePath(servicesCounselingRoutes.BASE_SHIPMENT_PPM_REVIEW_PATH, { moveCode, shipmentId }));
   };
 
-  const handleSubmit = (values) => {
+  const handleSubmit = () => {
     setErrorMessage(null);
     const ppmShipmentId = mtoShipment.ppmShipment.id;
 
-    const payload = {
-      certification_text: ppmSubmissionCertificationText,
-      signature: values.signature,
-      date: values.date,
-    };
-
-    submitPPMShipmentSignedCertification(ppmShipmentId, payload)
+    submitPPMShipmentSignedCertification(ppmShipmentId)
       .then(() => {
         setFlashMessage('PPM_SUBMITTED', 'success', 'You submitted documentation for review.', undefined, false);
         navigate(generatePath(servicesCounselingRoutes.BASE_MOVE_VIEW_PATH, { moveCode }));
@@ -63,33 +56,37 @@ const FinalCloseout = () => {
   };
 
   return (
-    <div className={classnames(ppmPageStyles.ppmPageStyle, styles.FinalCloseout)}>
-      <NotificationScrollToTop dependency={errorMessage} />
+    <div className={ppmPageStyles.tabContent}>
+      <div className={classnames(ppmPageStyles.container, styles.FinalCloseout)}>
+        <NotificationScrollToTop dependency={errorMessage} />
 
-      <GridContainer>
-        <Grid row>
-          <Grid col desktop={{ col: 8, offset: 2 }}>
-            <ShipmentTag shipmentType={shipmentTypes.PPM} />
+        <GridContainer className={ppmPageStyles.gridContainer}>
+          <Grid row>
+            <Grid col desktop={{ col: 8, offset: 2 }}>
+              <div className={ppmPageStyles.closeoutPageWrapper}>
+                <ShipmentTag shipmentType={shipmentTypes.PPM} />
 
-            <h1 data-testid="scCompletePPMHeader">Complete PPM</h1>
+                <h1 data-testid="scCompletePPMHeader">Complete PPM</h1>
 
-            {errorMessage && (
-              <Alert headingLevel="h4" slim type="error">
-                {errorMessage}
-              </Alert>
-            )}
+                {errorMessage && (
+                  <Alert headingLevel="h4" slim type="error">
+                    {errorMessage}
+                  </Alert>
+                )}
 
-            <FinalCloseoutForm
-              initialValues={initialValues}
-              mtoShipment={mtoShipment}
-              onBack={handleBack}
-              onSubmit={handleSubmit}
-              affiliation={order.agency}
-              selectedMove={selectedMove}
-            />
+                <FinalCloseoutForm
+                  initialValues={initialValues}
+                  mtoShipment={mtoShipment}
+                  onBack={handleBack}
+                  onSubmit={handleSubmit}
+                  affiliation={order.agency}
+                  selectedMove={selectedMove}
+                />
+              </div>
+            </Grid>
           </Grid>
-        </Grid>
-      </GridContainer>
+        </GridContainer>
+      </div>
     </div>
   );
 };
